+++ docs/changelog.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Improved accuracy of the [rule/duplicate](checks/rule/duplicate.md) check.
+- Fixed GitHub reporter trying to create pull request comments to unmodified lines - #1120.
 
 ## v0.65.1